+++ rtl/ibus/ibdr_dl11.vhd
@@ -1,6 +1,6 @@
--- $Id: ibdr_dl11.vhd 984 2018-01-02 20:56:27Z mueller $
+-- $Id: ibdr_dl11.vhd 1128 2019-04-07 13:12:47Z mueller $
 --
--- Copyright 2008-2011 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+-- Copyright 2008-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
 -- This program is free software; you may redistribute and/or modify it under
 -- the terms of the GNU General Public License as published by the Free
@@ -15,10 +15,10 @@
 -- Module Name:    ibdr_dl11 - syn
 -- Description:    ibus dev(rem): DL11-A/B
 --
--- Dependencies:   -
+-- Dependencies:   ib_rlim_slv
 -- Test bench:     -
 -- Target Devices: generic
--- Tool versions:  ise 8.2-14.7; viv 2014.4; ghdl 0.18-0.31
+-- Tool versions:  ise 8.2-14.7; viv 2017.2; ghdl 0.18-0.35
 --
 -- Synthesized (xst):
 -- Date         Rev  ise         Target      flop lutl lutm slic t peri
@@ -28,6 +28,9 @@
 --
 -- Revision History: 
 -- Date         Rev Version  Comment
+-- 2019-04-07  1127   1.3    for dl11_buf compat: xbuf.val in bit 15 and 8;
+--                           use rbuf instead xbuf for rdry reporting; remove
+--                           maintenance mode; use ib_rlim_slv; drop CE_USEC
 -- 2011-11-18   427   1.2.2  now numeric_std clean
 -- 2010-10-23   335   1.2.1  rename RRI_LAM->RB_LAM;
 -- 2010-10-17   333   1.2    use ibus V2 interface
@@ -56,9 +59,9 @@ entity ibdr_dl11 is                     -- ibus dev(rem): DL11-A/B
     IB_ADDR : slv16 := slv(to_unsigned(8#177560#,16)));
   port (
     CLK : in slbit;                     -- clock
-    CE_USEC : in slbit;                 -- usec pulse
     RESET : in slbit;                   -- system reset
     BRESET : in slbit;                  -- ibus reset
+    RLIM_CEV : in  slv7;                -- clock enable vector
     RB_LAM : out slbit;                 -- remote attention
     IB_MREQ : in ib_mreq_type;          -- ibus request
     IB_SRES : out ib_sres_type;         -- ibus response
@@ -80,12 +83,13 @@ architecture syn of ibdr_dl11 is
   constant rcsr_ibf_rdone : integer :=  7;
   constant rcsr_ibf_rie :   integer :=  6;
   
+  constant rbuf_ibf_rrdy :  integer := 15;
+  
   constant xcsr_ibf_xrdy :  integer :=  7;
   constant xcsr_ibf_xie :   integer :=  6;
-  constant xcsr_ibf_xmaint: integer :=  2;
 
-  constant xbuf_ibf_xval :  integer :=  8;
-  constant xbuf_ibf_rrdy :  integer :=  9;
+  constant xbuf_ibf_xval :  integer := 15;
+  constant xbuf_ibf_xval8 : integer :=  8;
 
   type regs_type is record              -- state registers
     ibsel : slbit;                      -- ibus select
@@ -95,11 +99,8 @@ architecture syn of ibdr_dl11 is
     rbuf : slv8;                        -- rbuf:
     rval : slbit;                       -- rx rbuf valid
     rintreq : slbit;                    -- rx interrupt request
-    rdlybsy : slbit;                    -- rx delay busy
-    rdlycnt : slv10;                    -- rx delay counter
     xrdy : slbit;                       -- xcsr: transmitter ready
     xie : slbit;                        -- xcsr: transmitter interrupt enable
-    xmaint : slbit;                     -- xcsr: maintenance mode
     xbuf : slv8;                        -- xbuf:
     xintreq : slbit;                    -- tx interrupt request
   end record regs_type;
@@ -109,28 +110,40 @@ architecture syn of ibdr_dl11 is
     (others=>'0'),                      -- rrlim
     '0','0',                            -- rdone, rie
     (others=>'0'),                      -- rbuf
-    '0','0','0',                        -- rval,rintreq,rdlybsy
-    (others=>'0'),                      -- rdlycnt
+    '0','0',                            -- rval,rintreq
     '1',                                -- xrdy !! is set !!
-    '0','0',                            -- xie,xmaint
+    '0',                                -- xie
     (others=>'0'),                      -- xbuf
     '0'                                 -- xintreq
   );
 
   signal R_REGS : regs_type := regs_init;
   signal N_REGS : regs_type := regs_init;
+  
+  signal RRLIM_START : slbit := '0';
+  signal RRLIM_BUSY  : slbit := '0';
 
 begin
   
+  RRLIM : ib_rlim_slv
+    port map (
+      CLK      => CLK,
+      RESET    => RESET,
+      RLIM_CEV => RLIM_CEV,
+      SEL      => R_REGS.rrlim,
+      START    => RRLIM_START,
+      STOP     => BRESET,
+      DONE     => open,
+      BUSY     => RRLIM_BUSY
+    );
+
   proc_regs: process (CLK)
   begin
     if rising_edge(CLK) then
       if BRESET = '1' then
         R_REGS <= regs_init;
         if RESET = '0' then               -- if RESET=0 we do just an ibus reset
-          R_REGS.rrlim   <= N_REGS.rrlim;   -- don't reset rx rate limit
-          R_REGS.rdlybsy <= N_REGS.rdlybsy; -- don't reset rx delay busy
-          R_REGS.rdlycnt <= N_REGS.rdlycnt; -- don't reset rx delay counter
+          R_REGS.rrlim <= N_REGS.rrlim;     -- keep RLIM flag
         end if;
       else
         R_REGS <= N_REGS;
@@ -138,7 +151,7 @@ begin
     end if;
   end process proc_regs;
 
-  proc_next : process (CE_USEC, R_REGS, IB_MREQ, EI_ACK_RX, EI_ACK_TX)
+  proc_next : process (R_REGS, IB_MREQ, EI_ACK_RX, EI_ACK_TX, RRLIM_BUSY)
     variable r : regs_type := regs_init;
     variable n : regs_type := regs_init;
     variable idout : slv16 := (others=>'0');
@@ -147,8 +160,7 @@ begin
     variable ibw0 : slbit := '0';
     variable ibw1 : slbit := '0';
     variable ilam : slbit := '0';
-    variable rdlystart : slbit := '0';
-    variable rdlyinit : slv10 := (others=>'0');
+    variable irrlimsta : slbit := '0';
   begin
 
     r := R_REGS;
@@ -160,7 +172,7 @@ begin
     ibw0  := IB_MREQ.we and IB_MREQ.be0;
     ibw1  := IB_MREQ.we and IB_MREQ.be1;
     ilam  := '0';
-    rdlystart := '0';
+    irrlimsta := '0';
       
     -- ibus address decoder
     n.ibsel := '0';
@@ -201,26 +213,17 @@ begin
           idout(r.rbuf'range)   := r.rbuf;
 
           if IB_MREQ.racc = '0' then    -- cpu ---------------------
-            if ibrd = '1' then
-              n.rdone   := '0';           -- clear DONE
+            if ibrd='1' and r.rdone='1' then
               n.rval    := '0';           -- clear rbuf valid
-              n.rintreq := '0';           -- clear pending interrupts
-              rdlystart := '1';           -- start rx delay counter
-              if r.xmaint = '0' then      -- if not in loop-back
-                ilam := '1';                -- request rb attention
-              end if;
+              irrlimsta := '1';           -- start rx timer
+              ilam := '1';                -- request rb attention
             end if;
 
           else                          -- rri ---------------------
+            idout(rbuf_ibf_rrdy) := not r.rval;
             if ibw0 = '1' then
               n.rbuf := IB_MREQ.din(n.rbuf'range);
               n.rval := '1';              -- set rbuf valid
-              if r.rdlybsy = '0' then     -- if rdly timer not running
-                n.rdone := '1';             -- set DONE
-                if r.rie = '1' then         -- if rx interrupt enabled
-                  n.rintreq := '1';           -- request interrupt
-                end if;
-              end if;
             end if;
           end if;
 
@@ -228,7 +231,6 @@ begin
 
           idout(xcsr_ibf_xrdy)  := r.xrdy;
           idout(xcsr_ibf_xie)   := r.xie;
-          idout(xcsr_ibf_xmaint):= r.xmaint;
 
           if IB_MREQ.racc = '0' then    -- cpu ---------------------
             if ibw0 = '1' then
@@ -240,7 +242,6 @@ begin
               else
                 n.xintreq := '0';
               end if;
-              n.xmaint := IB_MREQ.din(xcsr_ibf_xmaint);
             end if;
           end if;
           
@@ -251,17 +252,13 @@ begin
               n.xbuf := IB_MREQ.din(n.xbuf'range);
               n.xrdy := '0';
               n.xintreq := '0';
-              if r.xmaint = '0' then
-                ilam := '1';
-              end if;
+              ilam := '1';
             end if;
 
           else                          -- rri ---------------------
-            idout(r.xbuf'range)  := r.xbuf;
-            if r.xmaint = '0' then        -- if not in maintenace mode
-              idout(xbuf_ibf_xval) := not r.xrdy;
-              idout(xbuf_ibf_rrdy) := not r.rval;
-            end if;
+            idout(r.xbuf'range)   := r.xbuf;
+            idout(xbuf_ibf_xval)  := not r.xrdy;
+            idout(xbuf_ibf_xval8) := not r.xrdy;
             if ibrd = '1' then
               n.xrdy := '1';
               if r.xie = '1' then
@@ -273,59 +270,10 @@ begin
         when others => null;
       end case;
 
-    else                                -- if unselected handle loop-back
-      if r.xmaint = '1' and               -- if in maintenace mode
-          r.xrdy='0' and                  -- and transmit pending
-          r.rdone='0' and                 -- and receive buffer empty
-          r.rdlybsy='0' then              -- and rdly timer not running
-        n.rbuf  := r.xbuf;                  -- copy transmit to receive buffer
-        n.xrdy  := '1';                     -- mark transmit done
-        n.rdone := '1';                     -- make receive done
-        if r.rie = '1' then                 -- if rx interrupt enabled
-          n.rintreq := '1';                   -- request it
-        end if;
-        if r.xie = '1' then                 -- if tx interrupt enabled
-          n.xintreq := '1';                   -- request it
-        end if;
-      end if;
-        
     end if;    
 
     -- other state changes
-
-    rdlyinit := (others=>'0');
-    case r.rrlim is
-      when "000" => rdlyinit := "0000000000"; -- rlim=0 -> disabled
-      when "001" => rdlyinit := "0000000011"; -- rlim=1 -> delay by    3+ usec
-      when "010" => rdlyinit := "0000001111"; -- rlim=2 -> delay by   15+ usec
-      when "011" => rdlyinit := "0000111111"; -- rlim=3 -> delay by   63+ usec
-      when "100" => rdlyinit := "0001111111"; -- rlim=4 -> delay by  127+ usec
-      when "101" => rdlyinit := "0011111111"; -- rlim=5 -> delay by  255+ usec
-      when "110" => rdlyinit := "0111111111"; -- rlim=6 -> delay by  511+ usec
-      when "111" => rdlyinit := "1111111111"; -- rlim=7 -> delay by 1023+ usec
-      when others => null;
-    end case;
-    
-    if rdlystart = '1' then                 -- if rdly timer start requested
-      n.rdlycnt := rdlyinit;                  -- init counter
-      if r.rrlim /= "000" then                -- rate limiter enabled ?
-        n.rdlybsy := '1';                       -- set busy 
-      end if;
-    elsif CE_USEC = '1' then                -- if end-of-usec
-      n.rdlycnt := slv(unsigned(r.rdlycnt) - 1);   -- decrement
-      if r.rdlybsy='1' and                   -- if delay busy
-          unsigned(r.rdlycnt) = 0 then        --   and counter at zero
-        n.rdlybsy := '0';                       -- clear busy
-        if n.rval = '1' then                    -- if rbuf is valid or is set
-                                                --   valid this cycle (use n.!!)
-          n.rdone := '1';                         -- set DONE
-          if r.rie = '1' then                     -- if rx interrupt enabled
-            n.rintreq := '1';                       -- request interrupt 
-          end if;
-        end if;
-      end if;
-    end if;
-    
+  
     if EI_ACK_RX = '1' then
       n.rintreq := '0';
     end if;
@@ -333,7 +281,19 @@ begin
       n.xintreq := '0';
     end if;
 
+    if (RRLIM_BUSY or (not r.rval)) = '1' then -- busy or no data
+      n.rdone   := '0';                          -- clear done
+      n.rintreq := '0';                          -- clear pending interrupts
+    else                                       -- not busy and data valid
+      n.rdone := '1';                            -- clear done
+      if r.rdone='0' and r.rie='1' then          -- done going 0->1 and ie=1
+        n.rintreq := '1';                        -- request rx interrupt
+      end if;
+    end if;
+    
     N_REGS <= n;
+    
+    RRLIM_START <= irrlimsta;
 
     IB_SRES.dout <= idout;
     IB_SRES.ack  <= r.ibsel and ibreq;