+++ tools/oskit/rsx11mp-30_rp/rsx11mp-30_rp_boot.tcl
@@ -1,4 +1,4 @@
-# $Id: rsx11mp-30_rp_boot.tcl 835 2016-12-31 10:00:14Z mueller $
+# $Id: rsx11mp-30_rp_boot.tcl 1126 2019-04-06 17:37:40Z mueller $
 #
 # Setup file for RSX11-M+ V3.0 RP06 based system
 #
@@ -13,9 +13,10 @@
 rutil::dohook "preinithook"
 puts [rlw]
 
-# setup tt,lp (211bsd uses parity -> use 7 bit mode)
+# setup tt,lp,pp
 rw11::setup_tt "cpu0" to7bit 1
 rw11::setup_lp 
+rw11::setup_pp
 
 # mount disks
 cpu0rpa0 set type rp06