+++ tools/src/librw11/Rw11CntlDL11.hpp
@@ -1,6 +1,6 @@
-// $Id: Rw11CntlDL11.hpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11CntlDL11.hpp 1126 2019-04-06 17:37:40Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-04-06  1126   1.3    xbuf.val in msb; rrdy in rbuf (new iface)
 // 2017-05-14   897   1.2    add RcvChar(),TraceChar()
 // 2017-04-02   865   1.1.1  Dump(): add detail arg
 // 2014-12-29   623   1.1    adopt to Rlink V4 attn logic
@@ -23,7 +24,6 @@
 
 
 /*!
-  \file
   \brief   Declaration of class Rw11CntlDL11.
 */
 
@@ -67,14 +67,25 @@ namespace Retro {
       static const bool     kProbeInt = true;  //!< probe int active
       static const bool     kProbeRem = true;  //!< probr rem active
 
-      static const uint16_t kRCSR_M_RXRLIM = 0070000;
-      static const uint16_t kRCSR_V_RXRLIM = 12;
-      static const uint16_t kRCSR_B_RXRLIM = 007;
-      static const uint16_t kRCSR_M_RDONE  = kWBit07;
-      static const uint16_t kXCSR_M_XRDY   = kWBit07;
-      static const uint16_t kXBUF_M_RRDY   = kWBit09;
-      static const uint16_t kXBUF_M_XVAL   = kWBit08;
-      static const uint16_t kXBUF_M_XBUF   = 0xff;
+      static const uint16_t kRCSR_M_RLIM  = 0070000;
+      static const uint16_t kRCSR_V_RLIM  = 12;      //!< rcsr.rlim shift
+      static const uint16_t kRCSR_B_RLIM  = 007;     //!< rcsr.rlim bit mask
+      static const uint16_t kRCSR_V_TYPE  =  8;      //!< rcsr.type shift
+      static const uint16_t kRCSR_B_TYPE  = 0007;    //!< rcsr.type bit mask
+      static const uint16_t kRCSR_M_RDONE = kWBit07; //!< rcsr.rdone mask
+      static const uint16_t kRCSR_M_FCLR  = kWBit05; //!< rcsr.fclr mask    
+      static const uint16_t kRBUF_M_RRDY  = kWBit15; //!< rbuf.rrdy mask
+      static const uint16_t kRBUF_V_SIZE  =  8;      //!< rbuf.size shift
+      static const uint16_t kRBUF_B_SIZE  = 0177;    //!< rbuf.size bit mask
+      static const uint16_t kRBUF_M_BUF   = 0377;    //!< rbuf data mask
+  
+      static const uint16_t kXCSR_V_RLIM  = 12;      //!< xcsr.rlim shift 
+      static const uint16_t kXCSR_B_RLIM  = 007;     //!< xcsr.rlim bit mask
+      static const uint16_t kXCSR_M_XRDY  = kWBit07; //!< xcsr.xrdy mask
+      static const uint16_t kXBUF_M_VAL   = kWBit15; //!< xbuf.val mask
+      static const uint16_t kXBUF_V_SIZE  =  8;      //!< xbuf.size shift
+      static const uint16_t kXBUF_B_SIZE  = 0177;    //!< xbuf.size bit mask
+      static const uint16_t kXBUF_M_BUF   = 0xff;    //!< xbuf data mask
 
     protected:
       int           AttnHandler(RlinkServer::AttnArgs& args);
@@ -83,6 +94,7 @@ namespace Retro {
     
     protected:
       size_t        fPC_xbuf;               //!< PrimClist: xbuf index
+      size_t        fPC_rbuf;               //!< PrimClist: rbuf index
       uint16_t      fRxRlim;                //!< rx interrupt rate limit
   };