+++ tools/tcl/ibd_dl11/util.tcl
@@ -1,6 +1,6 @@
-# $Id: util.tcl 985 2018-01-03 08:59:40Z mueller $
+# $Id: util.tcl 1126 2019-04-06 17:37:40Z mueller $
 #
-# Copyright 2015- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+# Copyright 2015-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 #
 # This program is free software; you may redistribute and/or modify it under
 # the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 #
 #  Revision History:
 # Date         Rev Version  Comment
+# 2019-04-06  1126   1.1    updates for buffered dl11
 # 2015-12-26   719   1.0    Initial version
 #
 
@@ -27,12 +28,25 @@ namespace eval ibd_dl11 {
   # setup register descriptions for ibd_dl11 ---------------------------------
   #
 
-  regdsc RCSR  {done 7} {ie 6}
-  regdsc RRCSR {done 7} {ie 6} {rlim 14 3}
+  regdsc RCSR   {done 7} {ie 6}
+  regdsc RRCSR  {rlim 14 3} {type 10 3} {done 7} {ie 6} {fclr 5}
+  regdsc RRBUF  {rrdy 15} {size 14 7 "d"}
 
-  regdsc XCSR  {done 7} {ie 6} {maint 2}
+  regdsc XCSR   {done 7} {ie 6}
+  regdsc RXCSR  {rlim 14 3} {done 7} {ie 6}
+  regdsc RXBUF  {val 15} {size 14 7 "d"} {data 7 8 "o"}
 
   rw11util::regmap_add ibd_dl11 tt?.rcsr {l? RCSR r? RRCSR}
-  rw11util::regmap_add ibd_dl11 tt?.xcsr {?? XCSR}
+  rw11util::regmap_add ibd_dl11 tt?.rbuf {r? RRBUF}
+  rw11util::regmap_add ibd_dl11 tt?.xcsr {l? XCSR r? RXCSR}
+  rw11util::regmap_add ibd_dl11 tt?.xbuf {r? RXBUF}
+  
+  variable ANUM  1;             # 1st DL11
 
+  #
+  # setup: create controller with default attributes -------------------------
+  #
+  proc setup {{cpu "cpu0"}} {
+    return [rw11::setup_cntl $cpu "dl11" "tta"]
+  }
 }